# api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cat Activities Monitor API",
    "description": "\n    A comprehensive FastAPI-based system for monitoring cats using YOLO object detection and activity analysis.    ",
    "license": {
      "name": "MIT License",
      "url": "https://github.com/biancarosa/cat-activities-monitor/blob/main/LICENSE"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/feedback": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "List Feedback",
        "description": "List all submitted feedback for review.",
        "operationId": "list_feedback_feedback_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "feedback"
        ],
        "summary": "Submit Feedback",
        "description": "Submit feedback for a detection image to improve the model.",
        "operationId": "submit_feedback_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/feedback/{feedback_id}": {
      "get": {
        "tags": [
          "feedback"
        ],
        "summary": "Get Feedback Details",
        "description": "Get detailed feedback information.",
        "operationId": "get_feedback_details_feedback__feedback_id__get",
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feedback Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "feedback"
        ],
        "summary": "Delete Feedback",
        "description": "Delete feedback entry.",
        "operationId": "delete_feedback_feedback__feedback_id__delete",
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feedback Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BoundingBox": {
        "properties": {
          "x1": {
            "type": "number",
            "minimum": 0.0,
            "title": "X1",
            "description": "Left edge x-coordinate in pixels",
            "example": 100.5
          },
          "y1": {
            "type": "number",
            "minimum": 0.0,
            "title": "Y1",
            "description": "Top edge y-coordinate in pixels",
            "example": 150.2
          },
          "x2": {
            "type": "number",
            "minimum": 0.0,
            "title": "X2",
            "description": "Right edge x-coordinate in pixels",
            "example": 300.8
          },
          "y2": {
            "type": "number",
            "minimum": 0.0,
            "title": "Y2",
            "description": "Bottom edge y-coordinate in pixels",
            "example": 400.1
          },
          "width": {
            "type": "number",
            "minimum": 0.0,
            "title": "Width",
            "description": "Width of the bounding box in pixels",
            "example": 200.3
          },
          "height": {
            "type": "number",
            "minimum": 0.0,
            "title": "Height",
            "description": "Height of the bounding box in pixels",
            "example": 249.9
          }
        },
        "type": "object",
        "required": [
          "x1",
          "y1",
          "x2",
          "y2",
          "width",
          "height"
        ],
        "title": "BoundingBox",
        "description": "Standardized bounding box model for feedback.\n\nRepresents rectangular coordinates in pixel space with additional\nwidth and height for convenience."
      },
      "CatActivity": {
        "type": "string",
        "enum": [
          "unknown",
          "sitting",
          "lying",
          "standing",
          "moving",
          "eating",
          "playing",
          "sleeping",
          "grooming"
        ],
        "title": "CatActivity",
        "description": "Enumeration of detectable cat activities.\n\nThese activities are recognized through pose analysis and behavioral patterns."
      },
...
      "Detection": {
        "properties": {
          "class_id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Class Id",
            "description": "YOLO class ID for the detected object",
            "example": 15
          },
          "class_name": {
            "type": "string",
            "title": "Class Name",
            "description": "Human-readable name of the detected class",
            "example": "cat"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Detection confidence score between 0.0 and 1.0",
            "example": 0.85
          },
          "bounding_box": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Bounding Box",
            "description": "Bounding box coordinates in pixels",
            "example": {
              "height": 249.9,
              "width": 200.3,
              "x1": 100.5,
              "x2": 300.8,
              "y1": 150.2,
              "y2": 400.1
            }
          }
        },
        "type": "object",
        "required": [
          "class_id",
          "class_name",
          "confidence",
          "bounding_box"
        ],
        "title": "Detection",
        "description": "Represents a single object detection result from ML model.\n\nContains the detected object's class information, confidence score,\nand bounding box coordinates."
      },
      "FeedbackAnnotation": {
        "properties": {
          "class_id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Class Id",
            "description": "YOLO class ID for the annotated object",
            "example": 15
          },
          "class_name": {
            "type": "string",
            "title": "Class Name",
            "description": "Human-readable name of the annotated class",
            "example": "cat"
          },
          "bounding_box": {
            "$ref": "#/components/schemas/BoundingBox",
            "description": "Corrected or confirmed bounding box coordinates"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "User-provided confidence rating for this annotation",
            "example": 0.95
          },
          "cat_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Cat Name",
            "description": "Optional name for the identified cat",
            "example": "Whiskers"
          },
          "activity_feedback": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Feedback",
            "description": "User's textual feedback about the cat's activity",
            "example": "The cat appears to be playing with a toy mouse"
          },
          "correct_activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatActivity"
              },
              {
                "type": "null"
              }
            ],
            "description": "User's correction of the detected activity"
          },
          "activity_confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Confidence",
            "description": "User's confidence in their activity assessment",
            "example": 0.9
          }
        },
        "type": "object",
        "required": [
          "class_id",
          "class_name",
          "bounding_box"
        ],
        "title": "FeedbackAnnotation",
        "description": "User-provided annotation for training data.\n\nContains corrections, additions, or confirmations of detection results\nalong with optional cat identification and activity feedback."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageFeedback": {
        "properties": {
          "image_filename": {
            "type": "string",
            "title": "Image Filename",
            "description": "Name of the image file being annotated",
            "example": "camera1_20231201_143022_activity_detections.jpg"
          },
          "image_path": {
            "type": "string",
            "title": "Image Path",
            "description": "Full path to the image file",
            "example": "/detections/camera1_20231201_143022_activity_detections.jpg"
          },
          "original_detections": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "type": "array",
            "title": "Original Detections",
            "description": "Original detection results from the ML model"
          },
          "user_annotations": {
            "items": {
              "$ref": "#/components/schemas/FeedbackAnnotation"
            },
            "type": "array",
            "title": "User Annotations",
            "description": "User-provided corrections and annotations"
          },
          "feedback_type": {
            "type": "string",
            "pattern": "^(correction|addition|confirmation|rejection)$",
            "title": "Feedback Type",
            "description": "Type of feedback being provided",
            "example": "correction"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Additional notes or comments from the user",
            "example": "The second cat was missed by the detector - it's partially hidden behind the first cat"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "When this feedback was submitted",
            "example": "2023-12-01T14:30:22.123456"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "Identifier for the user providing feedback",
            "default": "anonymous",
            "example": "user123"
          }
        },
        "type": "object",
        "required": [
          "image_filename",
          "image_path",
          "original_detections",
          "user_annotations",
          "feedback_type",
          "timestamp"
        ],
        "title": "ImageFeedback",
        "description": "Feedback data for a single detection image.\n\nContains the original detection results along with user corrections,\nadditions, and annotations for model improvement."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}